requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
allure-python-commons-test==2.8.16
apipkg==1.5
appdirs==1.4.4
attrs==19.3.0
distlib==0.3.0
execnet==1.7.1
filelock==3.0.12
glob2==0.7
importlib-metadata==1.6.0
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
parse==1.15.0
parse-type==0.5.2
pluggy==0.13.1
py==1.8.1
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==5.4.2
pytest-bdd==3.3.0
pytest-forked==1.1.3
pytest-xdist==1.32.0
six==1.15.0
toml==0.10.1
tox==3.15.1
virtualenv==20.0.21
wcwidth==0.1.9
zipp==3.1.0
